runtests/requirements.txt
Django>=4.0
django-jsonfield-backport
django-extensions
ipdb
werkzeug
rest-framework-generic-relations
django-debug-toolbar
django-rest-framework
